Rework playRound() pseudocode to return result
Adjusted playRound() pseudocode to log the win/lose/tie strings rather than return
them, and created a new variable result which would hold a shorter
string that would present the result of the round when returned. I did
this in order to simplify future calls for this function when
constructing a score board.