Code the pseudocode, copy textContent in playRound
Converted the new button pseudocode into code, which fixed the issue
of the buttons not forming. The most major change was copying the
textContent changes of scoreBoard and declareWinner into the event
function. This allowed the scoreboards to be properly updated each
time the eventListener was triggered. A rather anti-climatic finish
to the project that halted me in my tracks on my last attempt at the
Odin Project. I essentially just used the code that was given to me to
and added a content declaration to it in order to solve my problems. It
didn't feel like I solved the problem in the planning stages of pseudo-
coding. C'est la vie.